Added Play Store app as requirement
Added to section 3.4: 
using the Google Play Store _App on an Android phone_

To make it explicit that the URL needs to be used by the Play Store app on an Android phone. As installing using the webversion of the Play Store is not supported by the Google Install Referrer API that is used. We use prefix `https://play.app.goo.gl/?link=` to enforce opening the link with Google Play Store. It is possible that applications still redirect to Play Store web. Firefox is a known exception. In this Browser users need to select that they want to open the url with the Play Store app